requirements.txt
streamlit==1.32.0 
pandas==2.2.0 
qrcode==7.4.2 
Pillow==10.2.0 
opencv-python-headless==4.9.0.80 
pyzbar==0.1.9 
Flask==3.0.0 
python-dotenv==1.0.0 
supabase==2.4.1
openpyxl==3.1.2 
plotly 
matplotlib 
seaborn 
